requirements.txt
telethon==1.42.0
python-dotenv==1.2.1
Pillow==12.1.0
asyncpg==0.31.0